factorize index write operation
