+++ meilisearch-http/src/data.rs
@@ -1,9 +1,11 @@
 use std::ops::Deref;
 use std::sync::Arc;
 
+use chrono::{DateTime, Utc};
 use indexmap::IndexMap;
 use meilisearch_core::{update, Database, DatabaseOptions};
-use serde::Deserialize;
+use rand::seq::SliceRandom;
+use serde::{Deserialize, Serialize};
 use serde_json::Value;
 use sha2::Digest;
 
@@ -20,9 +22,44 @@ pub struct UpdateDocumentsQuery {
     primary_key: Option<String>,
 }
 
+#[derive(Debug, Serialize)]
+#[serde(rename_all = "camelCase")]
+pub struct IndexResponse {
+    pub name: String,
+    pub uid: String,
+    pub created_at: DateTime<Utc>,
+    pub updated_at: DateTime<Utc>,
+    pub primary_key: Option<String>,
+}
+
+#[derive(Debug, Deserialize)]
+#[serde(rename_all = "camelCase", deny_unknown_fields)]
+pub struct IndexCreateRequest {
+    name: Option<String>,
+    uid: Option<String>,
+    primary_key: Option<String>,
+}
+
 #[derive(Clone)]
 pub struct Data {
     inner: Arc<DataInner>,
+}
+
+#[derive(Debug, Deserialize)]
+#[serde(rename_all = "camelCase", deny_unknown_fields)]
+pub struct UpdateIndexRequest {
+    name: Option<String>,
+    primary_key: Option<String>,
+}
+
+#[derive(Debug, Serialize)]
+#[serde(rename_all = "camelCase")]
+pub struct UpdateIndexResponse {
+    name: String,
+    uid: String,
+    created_at: DateTime<Utc>,
+    updated_at: DateTime<Utc>,
+    primary_key: Option<String>,
 }
 
 impl Data {
@@ -109,6 +146,132 @@ impl Data {
 
         Ok(IndexUpdateResponse::with_id(update_id))
     }
+
+    pub fn create_index(&self, body: IndexCreateRequest) -> Result<IndexResponse, ResponseError> {
+        if let (None, None) = (body.name.clone(), body.uid.clone()) {
+            return Err(Error::bad_request("Index creation must have an uid").into());
+        }
+
+        let uid = match &body.uid {
+            Some(uid) => {
+                if uid
+                    .chars()
+                    .all(|x| x.is_ascii_alphanumeric() || x == '-' || x == '_')
+                {
+                    uid.to_owned()
+                } else {
+                    return Err(Error::InvalidIndexUid.into());
+                }
+            }
+            None => loop {
+                let uid = generate_uid();
+                if self.db.open_index(&uid).is_none() {
+                    break uid;
+                }
+            },
+        };
+
+        let created_index = self.db.create_index(&uid).map_err(|e| match e {
+            meilisearch_core::Error::IndexAlreadyExists => e.into(),
+            _ => ResponseError::from(Error::create_index(e)),
+        })?;
+
+        let index_response = self.db.main_write::<_, _, ResponseError>(|mut writer| {
+            let name = body.name.as_ref().unwrap_or(&uid);
+            created_index.main.put_name(&mut writer, name)?;
+
+            let created_at = created_index
+                .main
+                .created_at(&writer)?
+                .ok_or(Error::internal("Impossible to read created at"))?;
+
+            let updated_at = created_index
+                .main
+                .updated_at(&writer)?
+                .ok_or(Error::internal("Impossible to read updated at"))?;
+
+            if let Some(id) = body.primary_key.clone() {
+                if let Some(mut schema) = created_index.main.schema(&writer)? {
+                    schema.set_primary_key(&id).map_err(Error::bad_request)?;
+                    created_index.main.put_schema(&mut writer, &schema)?;
+                }
+            }
+            let index_response = IndexResponse {
+                name: name.to_string(),
+                uid,
+                created_at,
+                updated_at,
+                primary_key: body.primary_key.clone(),
+            };
+            Ok(index_response)
+        })?;
+
+        Ok(index_response)
+    }
+
+    pub fn update_index(
+        &self,
+        path: IndexParam,
+        body: IndexCreateRequest,
+    ) -> Result<IndexResponse, ResponseError> {
+        let index = self
+            .db
+            .open_index(&path.index_uid)
+            .ok_or(Error::index_not_found(&path.index_uid))?;
+
+        self.db.main_write::<_, _, ResponseError>(|writer| {
+            if let Some(name) = &body.name {
+                index.main.put_name(writer, name)?;
+            }
+
+            if let Some(id) = body.primary_key.clone() {
+                if let Some(mut schema) = index.main.schema(writer)? {
+                    schema.set_primary_key(&id)?;
+                    index.main.put_schema(writer, &schema)?;
+                }
+            }
+            index.main.put_updated_at(writer)?;
+            Ok(())
+        })?;
+
+        let reader = self.db.main_read_txn()?;
+        let name = index
+            .main
+            .name(&reader)?
+            .ok_or(Error::internal("Impossible to get the name of an index"))?;
+        let created_at = index.main.created_at(&reader)?.ok_or(Error::internal(
+            "Impossible to get the create date of an index",
+        ))?;
+        let updated_at = index.main.updated_at(&reader)?.ok_or(Error::internal(
+            "Impossible to get the last update date of an index",
+        ))?;
+
+        let primary_key = match index.main.schema(&reader) {
+            Ok(Some(schema)) => match schema.primary_key() {
+                Some(primary_key) => Some(primary_key.to_owned()),
+                None => None,
+            },
+            _ => None,
+        };
+
+        let index_response = IndexResponse {
+            name,
+            uid: path.index_uid.clone(),
+            created_at,
+            updated_at,
+            primary_key,
+        };
+
+        Ok(index_response)
+    }
+
+    pub fn delete_index(&self, path: IndexParam) -> Result<(), ResponseError> {
+        if self.db.delete_index(&path.index_uid)? {
+            Ok(())
+        } else {
+            Err(Error::index_not_found(&path.index_uid).into())
+        }
+    }
 }
 
 fn find_primary_key(document: &IndexMap<String, Value>) -> Option<String> {
@@ -118,6 +281,15 @@ fn find_primary_key(document: &IndexMap<String, Value>) -> Option<String> {
         }
     }
     None
+}
+
+fn generate_uid() -> String {
+    let mut rng = rand::thread_rng();
+    let sample = b"abcdefghijklmnopqrstuvwxyz0123456789";
+    sample
+        .choose_multiple(&mut rng, 8)
+        .map(|c| *c as char)
+        .collect()
 }
 
 impl Deref for Data {
